+++ test/Main.hs
@@ -5,6 +5,7 @@ import Test.Tasty
 import qualified Test.IR.Pretty as Pretty
 import qualified Test.IR.Builder as Builder
 import qualified Test.EDSL.Ops as EDSLOps
+import qualified Test.ModuleBuilder as ModuleBuilder
 import qualified Test.Runtime.EndToEnd as EndToEnd
 import qualified Test.Runtime.EndToEndArithmetic as Arith
 import qualified Test.Runtime.EndToEndShape as Shape
@@ -13,6 +14,7 @@ import qualified Test.Runtime.EndToEndNN as NN
 import qualified Test.Runtime.EndToEndReductions as Reductions
 import qualified Test.Runtime.EndToEndDataMovement as DataMovement
 import qualified Test.Runtime.EndToEndMultiValue as MultiValue
+import qualified Test.Runtime.EndToEndSession as Session
 import qualified Test.Runtime.Buffer as Buffer
 import qualified Test.Runtime.Async as Async
 import qualified Test.Runtime.Errors as Errors
@@ -44,6 +46,7 @@ main = do
         , Reductions.tests
         , DataMovement.tests
         , MultiValue.tests
+        , Session.tests
         , Buffer.tests
         , Async.tests
         , Errors.tests